Prueba de la carga de animes 2

+++ app/src/main/java/com/wincest/animedb/Anime.java
@@ -37,7 +37,7 @@ public class Anime
         return episodes;
     }
 
-    public float getRating()
+    public double getRating()
     {
         return rating;
     }
